applied patch from Dejan Bosanac for AMQ-1227
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@549845 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-web/src/main/java/org/apache/activemq/web/WebClient.java
@@ -205,7 +205,7 @@ public class WebClient implements HttpSessionActivationListener, HttpSessionBind
         }
     }
 
-    public void send(Destination destination, Message message, boolean persistent, int priority, int timeToLive) throws JMSException {
+    public void send(Destination destination, Message message, boolean persistent, int priority, long timeToLive) throws JMSException {
         int deliveryMode = persistent ? DeliveryMode.PERSISTENT : DeliveryMode.NON_PERSISTENT;
         getProducer().send(destination, message, deliveryMode, priority, timeToLive);
         if (log.isDebugEnabled()) {
